// afm-frontend/src/app/core/models/future-event-request.model.ts
export class FutureEventRequestModel {
  // event_name: string;
  // latitude: string;
  // longitude: string;
  // address: string;
  // zip_code: string;
  // event_date: string;
  // event_time_slot: string;
  // message: string;

  event_type: string;
  longitude: string;
  latitude: string;
  email_id: string;
  address: string;
  event_date: string;
  zip_code: string;
  message: string;
}